Finished breaking out repositories into partial class
