+++ .vscode/csharp.code-snippets
@@ -1,0 +1,13 @@
+{
+    // Place your snippets for csharp here. Each snippet is defined under a snippet name and has a prefix, body and
+    // description. The prefix is what is used to trigger the snippet and the body will be expanded and inserted. Possible variables are:
+    // $1, $2 for tab stops, $0 for the final cursor position, and ${1:label}, ${2:another} for placeholders. Placeholders with the
+    // same ids are connected.
+    // Snippets can be easily generated or modified using https://snippet-generator.app/
+
+    "C# Region": {
+        "prefix": "cs-region",
+        "body": ["#region $1", "", "#endregion $1"],
+        "description": "C# region"
+    }
+}
